Upgrade and simplify React-Native
Removes unnecessary files and classes that were added as part of the
create-react-native-app bundle. This works by integrating React-Native
into a new (empty) Android project, and then re-introducing the digi.me
SDKs so that the demo React-Native app can use the Consent Access platform.

This also updates Android to generate a build.properties file which contains
the applications details and credentials that are required for use Consent
Access, this is created from the `configure-native` yarn command.

`yarn compile` now compiles all Typescript to JavaScript and copies all
required assets to a single ./dist/ folder.