Extend NativeBridge to support more events, use singleton for dispatching
We were experiencing intermittent issues with the original code, sometimes receiving the error:

Bridge is not set. This is probably because you've explicitly synthesized the bridge in NativeBridge, even though it's inherited from RCTEventEmitter.

As proposed in this issue https://github.com/facebook/react-native/issues/15421#issuecomment-335346159, solution is to use a singleton of NativeBridge, and then dispatch events via it.